[clean]supabase_functionの関数を役割ごとに分割
